feat: marker mode cursor and readiness ring
