+++ src/index.css
@@ -161,6 +161,14 @@ body {
   }
 }
 
+@utility waveform-armed {
+  cursor: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='22' height='28'%3E%3Cpath d='M11 2C7 2 4 5 4 9c0 5 7 16 7 16s7-11 7-16c0-4-3-7-7-7z' fill='%23f5a623' stroke='white' stroke-width='1.6'/%3E%3Ccircle cx='11' cy='9' r='2.6' fill='white'/%3E%3C/svg%3E")
+    11 26, crosshair;
+  box-shadow: inset 0 0 0 1px color-mix(in srgb, var(--color-composer-warning) 18%, transparent), inset 0 0 20px -2px
+    color-mix(in srgb, var(--color-composer-warning) 18%, transparent);
+  transition: box-shadow 180ms ease-out;
+}
+
 /* Timeline panning mode */
 @layer components {
   .is-panning,
